@ebozduman 2025-12-12
Now, we have format check failure.
Unfortunately, the logs do not give useful information about the failure.